+++ src/utils/nav_data_sync/nav_data_sync.h
@@ -29,7 +29,7 @@ typedef struct {
   s32 TOW_residual_ns; /**< Residual to TOW_ms [ns] */
   s8 bit_polarity;
   u16 glo_orbit_slot;
-  u16 read_index;
+  u8 read_index;
   health_t health;
   bool valid;
   decode_sync_flags_t sync_flags;
